Add more visual information to PR/Issue/Branch.
