+++ src/main/scala/gitbucket/core/util/TextAvatarUtil.scala
@@ -13,32 +13,12 @@ object TextAvatarUtil {
   private val fontSize = 180
   private val roundSize = 60
   private val shadowSize = 20
+
   private val bgSaturation = 0.68f
   private val bgBlightness = 0.73f
   private val shadowBlightness = 0.23f
   private val font = new Font(Font.SANS_SERIF, Font.PLAIN, fontSize)
   private val transparent = new Color(0, 0, 0, 0)
-
-  // https://www.w3.org/TR/2008/REC-WCAG20-20081211/#relativeluminancedef
-  private def relativeLuminance(c: Color): Double = {
-    val rgb = Seq(c.getRed, c.getGreen, c.getBlue).map{_/255.0}.map{x => if (x <= 0.03928) x / 12.92 else math.pow((x + 0.055) / 1.055, 2.4)}
-    0.2126 * rgb(0) + 0.7152 * rgb(1) + 0.0722 * rgb(2)
-  }
-
-  // https://www.w3.org/TR/2008/REC-WCAG20-20081211/#contrast-ratiodef
-  private def contrastRatio(c1: Color, c2: Color): Double = {
-    val l1 = relativeLuminance(c1)
-    val l2 = relativeLuminance(c2)
-    if (l1 > l2) (l1 + 0.05) / (l2 + 0.05) else (l2 + 0.05) / (l1 + 0.05)
-  }
-
-  private def goodContrastColor(base: Color, c1: Color, c2: Color): Color = {
-    if (contrastRatio(base, c1) > contrastRatio(base, c2)) c1 else c2
-  }
-
-  private def strToHue(text: String): Float = {
-    Integer.parseInt(StringUtil.md5(text).substring(0, 2), 16) / 256f
-  }
 
   private def getCenterToDraw(drawText: String, font: Font, w: Int, h: Int): (Int, Int) = {
     val context = new FontRenderContext(new AffineTransform(), true, true)
@@ -78,9 +58,7 @@ object TextAvatarUtil {
   def textAvatar(nameText: String): Option[Array[Byte]] = {
     val drawText = nameText.substring(0, 1)
 
-    val bgHue = strToHue(nameText)
-    val bgColor = Color.getHSBColor(bgHue, bgSaturation, bgBlightness)
-    val fgColor = goodContrastColor(bgColor, Color.BLACK, Color.WHITE)
+    val (bgColor, fgColor) = ColorUtil.getStrandomColors(nameText, bgSaturation, bgBlightness)
 
     val font = new Font(Font.SANS_SERIF, Font.PLAIN, fontSize)
     if (font.canDisplayUpTo(drawText) == -1) Some(textImage(drawText, bgColor, fgColor)) else None
@@ -117,10 +95,8 @@ object TextAvatarUtil {
   def textGroupAvatar(nameText: String): Option[Array[Byte]] = {
     val drawText = nameText.substring(0, 1)
 
-    val bgHue = strToHue(nameText)
-    val bgColor = Color.getHSBColor(bgHue, bgSaturation, bgBlightness)
-    val fgColor = goodContrastColor(bgColor, Color.BLACK, Color.WHITE)
-    val shadowColor = Color.getHSBColor(bgHue, bgSaturation, shadowBlightness)
+    val (bgColor, fgColor) = ColorUtil.getStrandomColors(nameText, bgSaturation, bgBlightness)
+    val shadowColor = bgColor.darker()
 
     if (font.canDisplayUpTo(drawText) == -1) Some(textGroupImage(drawText, bgColor, fgColor, shadowColor)) else None
   }
